refactor: Header를 주로 사용하는 측인 layout/[lng]에 위치
