+++ packages/web/src/app/[lng]/(auth)/memos/_components/index.ts
@@ -1,4 +1,3 @@
-export { default as Header } from './Header';
 export { default as MemoDialog } from './MemoDialog';
 export { default as MemoSidebar } from './MemoSidebar';
 export { default as MemoSidebarTrigger } from './MemoSidebarTrigger';
