+++ packages/web/src/app/[lng]/_components/index.ts
@@ -1,0 +1,1 @@
+export { default as Header } from './Header';
